feat: [PLG-1164]: updating pr checks issues
