+++ src/modules/33-auth-settings/components/Subscription/BillingInfo/BillingInfo.tsx
@@ -17,9 +17,9 @@ import { getErrorMessage } from '@auth-settings/utils'
 import type { AccountPathProps } from '@common/interfaces/RouteInterfaces'
 import { ContainerSpinner } from '@common/components/ContainerSpinner/ContainerSpinner'
 import type { StateByCountryMap } from '@common/hooks/useRegionList'
+import type { Module } from 'framework/types/ModuleName'
 import BillingContact, { InitialBillingInfo } from './BillingContact'
 import { Header } from '../Header'
-import type { Module } from 'framework/types/ModuleName'
 import { PLAN_TYPES } from '../subscriptionUtils'
 import css from './BillingInfo.module.scss'
