+++ src/modules/33-auth-settings/components/Subscription/PricePreview/PricePreview.module.scss
@@ -10,7 +10,7 @@
 }
 .footerStyle {
   position: relative;
-  top: 120px;
+  top: 30px;
 }
 
 .parentTimeChild1 {
