+++ src/modules/33-auth-settings/components/Subscription/PricePreview/SubscriptionDetails.tsx
@@ -91,6 +91,7 @@ const SubscriptionDetails: React.FC<SubscriptionDetailsProps> = ({
       {premiumSupport && <PremiumSupportLine premiumSupportAmount={premiumSupportAmount} />}
       {<TotalAmount totalAmount={totalAmount || 0} />}
       {!isNil(subscriptionDetails.taxAmount) && <TaxLine taxAmount={subscriptionDetails.taxAmount} />}
+
       {isNil(subscriptionDetails.taxAmount) && (
         <Text font={{ size: 'xsmall' }}>{getString('authSettings.salesTax')}</Text>
       )}
